requirements.txt
Django==1.6.3
Markdown==2.4
MySQL-python==1.2.5
django-filter==0.7
djangorestframework==2.3.13
wsgiref==0.1.2
